+++ YGGG/YGGG/BookmarkTableViewController.swift
@@ -101,58 +101,3 @@ class BookmarkTableViewController: UIViewController, UITableViewDataSource, UITa
 
 }
 
-class CustomSearchBar: UISearchBar {
-    
-    var infoButton: UIButton = {
-        let infoButton = UIButton(type: .infoLight)
-        infoButton.setImage(UIImage(systemName: "info.circle"), for: .normal)
-        return infoButton
-    }()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupInfoButton()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setupInfoButton() {
-        infoButton.translatesAutoresizingMaskIntoConstraints = false
-        self.addSubview(infoButton)
-        
-        NSLayoutConstraint.activate([
-            infoButton.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -10),
-            infoButton.topAnchor.constraint(equalTo: self.topAnchor, constant: 5),
-            infoButton.widthAnchor.constraint(equalToConstant: 26),
-            infoButton.heightAnchor.constraint(equalToConstant: 26)
-        ])
-    }
-}
-
-class SearchInfoViewController: UIViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        view.backgroundColor = .white
-        let label = UILabel()
-        label.numberOfLines = 3
-        label.text = "검색 \n관련 \n설명"
-        label.textColor = .systemGray
-        label.font = UIFont.systemFont(ofSize: 12)
-        
-        view.addSubview(label)
-        label.translatesAutoresizingMaskIntoConstraints = false
-        
-        NSLayoutConstraint.activate([
-            label.topAnchor.constraint(equalTo: view.topAnchor, constant: 10),
-            label.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 10),
-            label.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: 10),
-            label.bottomAnchor.constraint(equalTo: view.bottomAnchor, constant: 10),
-        ])
-        
-        self.preferredContentSize = CGSize(width: 200, height: 200)
-    }
-}
-
